requirements.txt
torch
transformers
faiss-cpu
sentence-transformers
streamlit
watchdog